Honor redirect URLs upon logging in and out.
The plugin does authorize the user with the Identity Provider but fails to actually load the YOURLS `/admin` page after logging in. (On my system, only a blank page is displayed; one needs to manually refresh the browser in order to proceed.)

No signout page is displayed, either, because the log-out link `YOURLS_SITE` is not a webpage, whereas `YOURLS_SITE/admin` is.

This PR attempts to fix both.

Thank you for a very useful plugin.